Add trapezoid window function
Eric J specifically requested a rectangular window0. There did not seem to
be a workable solution in the scipy or numpy signal windows for the versions of
those packages we're using. The `boxcar` window does not include
endpoints of zero; the hamming and hanning windows have no tunable
parameters to adjust the window shape.