TST : Skip Volume Viewer tests on Travis
this is because we use a newer version of VTK and mayavi, which need a
newer version of OpenGL, which isn't available on Travis CI at the
moment.

This leads to tracebacks and failures on CI that are related to the
environment and not the package.

For now, we skip these tests on CI but run the tests locally using the
etstool utility test_clean

	modified:   .travis.yml
	modified:   ensemble/volren/tests/test_volume_renderer.py